+++ client/src/lifecycleManager.js
@@ -559,5 +559,19 @@ class LifecycleManager extends LuigiClientBase {
       anchor
     });
   }
+
+  /**
+   * This function allows you to change node labels within the same {@link navigation-advanced.md#view-groups view group}, e.g. in your node config: `label: 'my Node {viewGroupData.vg1}'`.
+   * @since 2.2.0
+   * @param {Object} data a data object containing the view group name and desired label
+   * @memberof Lifecycle
+   * @example LuigiClient.setViewGroupData({'vg1':' Luigi rocks!'})
+   */
+  setViewGroupData(data) {
+    helpers.sendPostMessageToLuigiCore({
+      msg: 'luigi.setVGData',
+      data
+    });
+  }
 }
 export const lifecycleManager = new LifecycleManager();