+++ core/src/App.svelte
@@ -246,18 +246,26 @@
     return paths.includes(removeQueryParams(routePath));
   };
 
+  /**
+   * Clears the dirty state when dirty state promise resolves and dirty state is not needed anymore
+   * @param source used for drawers/modals/split view wc and iframe when source is needed to differentiate which mf is affected
+   */
+  const clearDirtyState = (source) => {
+    if (unsavedChanges && unsavedChanges.dirtySet) {
+      if (source) {
+        unsavedChanges.dirtySet.delete(source);
+      } else {
+        unsavedChanges.dirtySet.clear();
+      }
+    }
+  };
+
   const getUnsavedChangesModalPromise = (source) => {
     return new Promise((resolve, reject) => {
       if (shouldShowUnsavedChangesModal(source)) {
         showUnsavedChangesModal().then(
           () => {
-            if (unsavedChanges && unsavedChanges.dirtySet) {
-              if (source) {
-                unsavedChanges.dirtySet.delete(source);
-              } else {
-                unsavedChanges.dirtySet.clear();
-              }
-            }
+            clearDirtyState();
             resolve();
           },
           () => {
@@ -503,23 +511,26 @@
 
   const handleNavClick = (event) => {
     const node = event.detail.node;
-    getUnsavedChangesModalPromise().then(() => {
-      closeLeftNav();
-      if (node.openNodeInModal) {
-        const route = RoutingHelpers.buildRoute(node, `/${node.pathSegment}`);
-        openViewInModal(
-          route,
-          node.openNodeInModal === true ? {} : node.openNodeInModal
-        );
-      } else if (node.drawer) {
-        const route = RoutingHelpers.buildRoute(node, `/${node.pathSegment}`);
-        node.drawer.isDrawer = true;
-        openViewInDrawer(route, node.drawer);
-      } else {
-        getComponentWrapper().set({ isNavigationSyncEnabled: true });
-        Routing.handleRouteClick(node, getComponentWrapper());
-      }
-    });
+    getUnsavedChangesModalPromise().then(
+      () => {
+        closeLeftNav();
+        if (node.openNodeInModal) {
+          const route = RoutingHelpers.buildRoute(node, `/${node.pathSegment}`);
+          openViewInModal(
+            route,
+            node.openNodeInModal === true ? {} : node.openNodeInModal
+          );
+        } else if (node.drawer) {
+          const route = RoutingHelpers.buildRoute(node, `/${node.pathSegment}`);
+          node.drawer.isDrawer = true;
+          openViewInDrawer(route, node.drawer);
+        } else {
+          getComponentWrapper().set({ isNavigationSyncEnabled: true });
+          Routing.handleRouteClick(node, getComponentWrapper());
+        }
+      },
+      () => {}
+    );
   };
 
   const onResizeTabNav = () => {
@@ -898,6 +909,12 @@
     });
   };
 
+  export const getDirtyStatus = () => {
+    return unsavedChanges.dirtySet
+      ? unsavedChanges.dirtySet.size > 0
+      : unsavedChanges.isDirty;
+  };
+
   setContext('getUnsavedChangesModalPromise', getUnsavedChangesModalPromise);
 
   //// MICRO-FRONTEND MODAL
@@ -997,7 +1014,8 @@
         () => {
           resetModalData(index, isClosedInternal);
           rp && rp.doResolve(goBackContext);
-        }
+        },
+        () => {}
       );
     } else if (targetModal && targetModal.modalWC) {
       resetModalData(index, isClosedInternal);
@@ -1084,13 +1102,19 @@
     ) {
       try {
         if (drawerIframe) {
-          getUnsavedChangesModalPromise(drawerIframe.contentWindow).then(() => {
-            resetMicrofrontendDrawerData();
-          });
+          getUnsavedChangesModalPromise(drawerIframe.contentWindow).then(
+            () => {
+              resetMicrofrontendDrawerData();
+            },
+            () => {}
+          );
         } else if (drawerWC) {
-          getUnsavedChangesModalPromise().then(() => {
-            resetMicrofrontendDrawerData();
-          });
+          getUnsavedChangesModalPromise().then(
+            () => {
+              resetMicrofrontendDrawerData();
+            },
+            () => {}
+          );
         }
         IframeHelpers.getCurrentMicrofrontendIframe().removeAttribute('style');
       } catch (e) {
@@ -1566,21 +1590,24 @@
         } else {
           // go back: context from the view
           if (preservedViews && preservedViews.length > 0) {
-            getUnsavedChangesModalPromise().then(() => {
-              // remove current active iframe and data
-              Iframe.setActiveIframeToPrevious(node);
-              const previousActiveIframeData = preservedViews.pop();
-              // set new active iframe and preservedViews
-              config.iframe = Iframe.getActiveIframe(node);
-              isNavigateBack = true;
-              preservedViews = preservedViews;
-              goBackContext = _goBackContext;
-              // TODO: check if getNavigationPath or history pop to update hash / path
-              handleNavigation(
-                { params: { link: previousActiveIframeData.path } },
-                config
-              );
-            });
+            getUnsavedChangesModalPromise().then(
+              () => {
+                // remove current active iframe and data
+                Iframe.setActiveIframeToPrevious(node);
+                const previousActiveIframeData = preservedViews.pop();
+                // set new active iframe and preservedViews
+                config.iframe = Iframe.getActiveIframe(node);
+                isNavigateBack = true;
+                preservedViews = preservedViews;
+                goBackContext = _goBackContext;
+                // TODO: check if getNavigationPath or history pop to update hash / path
+                handleNavigation(
+                  { params: { link: previousActiveIframeData.path } },
+                  config
+                );
+              },
+              () => {}
+            );
           } else {
             if (_goBackContext) {
               console.warn(
@@ -1779,6 +1806,16 @@
         const { anchor } = e.data;
         LuigiRouting.setAnchor(anchor);
       }
+
+      if ('luigi.setVGData' === e.data.msg) {
+        const vgData = e.data;
+        const vg = RoutingHelpers.findViewGroup(iframe.luigi.currentNode);
+        if (vg) {
+          const vgSettings = Iframe.getViewGroupSettings(vg);
+          vgSettings._liveCustomData = vgData.data;
+          LuigiConfig.configChanged('navigation.nodes');
+        }
+      }
     });
 
     // listeners are not automatically removed — cancel
@@ -1890,8 +1927,7 @@
   });
 
   beforeUpdate(() => {
-    breadcrumbsEnabled =
-      GenericHelpers.requestExperimentalFeature('breadcrumbs');
+    breadcrumbsEnabled = LuigiConfig.getConfigValue('navigation.breadcrumbs');
     searchProvider = LuigiConfig.getConfigValue('globalSearch.searchProvider');
     configTag = LuigiConfig.getConfigValue('tag');
     isHeaderDisabled = LuigiConfig.getConfigValue('settings.header.disabled');
@@ -1919,8 +1955,8 @@
         nodepath={modalItem.mfModal.nodepath}
         modalIndex={index}
         on:close={() => closeModal(index, true)}
-        on:iframeCreated={event => modalIframeCreated(event, index)}
-        on:wcCreated={event => modalWCCreated(event, index)}
+        on:iframeCreated={(event) => modalIframeCreated(event, index)}
+        on:wcCreated={(event) => modalWCCreated(event, index)}
         {disableBackdrop}
       />
     {/if}
@@ -1983,14 +2019,14 @@
       aria-label="Loading"
     >
       <div
-        class="fd-busy-indicator--m"
+        class="fd-busy-indicator fd-busy-indicator--m"
         aria-hidden="false"
         aria-label="Loading"
         data-testid="luigi-loading-spinner"
       >
-        <div class="fd-busy-indicator--circle-0" />
-        <div class="fd-busy-indicator--circle-1" />
-        <div class="fd-busy-indicator--circle-2" />
+        <div class="fd-busy-indicator__circle" />
+        <div class="fd-busy-indicator__circle" />
+        <div class="fd-busy-indicator__circle" />
       </div>
     </div>
   {/if}
@@ -2188,7 +2224,20 @@
 
   :global(.iframeContainer.iframeContainerTabNav) {
     top: calc(
-      var(--luigi__shellbar--height) + var(--luigi__horizontal-nav--height)
+      var(--luigi__shellbar--height) +
+        var(
+          --luigi__horizontal-nav--live-height,
+          var(--luigi__horizontal-nav--height)
+        )
+    );
+  }
+
+  :global(.no-top-nav .iframeContainer.iframeContainerTabNav) {
+    top: calc(
+      var(
+        --luigi__horizontal-nav--live-height,
+        var(--luigi__horizontal-nav--height)
+      )
     );
   }